Add modal to select bus capacity
